+++ models/nfs_inode.go
@@ -67,7 +67,7 @@ type NfsInode struct {
 	SnapshotNum *int32 `json:"snapshot_num"`
 
 	// unique logical size
-	UniqueLogicalSize *int64 `json:"unique_logical_size,omitempty"`
+	UniqueLogicalSize *float64 `json:"unique_logical_size,omitempty"`
 
 	// unique size
 	// Required: true
