+++ models/vm.go
@@ -195,7 +195,7 @@ type VM struct {
 	Status *VMStatus `json:"status"`
 
 	// unique logical size
-	UniqueLogicalSize *int64 `json:"unique_logical_size,omitempty"`
+	UniqueLogicalSize *float64 `json:"unique_logical_size,omitempty"`
 
 	// unique size
 	UniqueSize *int64 `json:"unique_size,omitempty"`
