+++ models/vm_volume.go
@@ -82,7 +82,7 @@ type VMVolume struct {
 	Type *VMVolumeType `json:"type,omitempty"`
 
 	// unique logical size
-	UniqueLogicalSize *int64 `json:"unique_logical_size,omitempty"`
+	UniqueLogicalSize *float64 `json:"unique_logical_size,omitempty"`
 
 	// unique size
 	UniqueSize *int64 `json:"unique_size,omitempty"`
